Add methods to hash password and add user to database
